fix: use Double return type in GameRendererMixin for 1.20.1
GameRenderer.getFov returns double in 1.20.1, not float.
ClassCastException crash when using spyglass.

+++ src/main/java/com/spyglasshud/mixin/GameRendererMixin.java
@@ -13,11 +13,11 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 public class GameRendererMixin {
 
     @Inject(method = "getFov", at = @At("RETURN"), cancellable = true)
-    private void modifySpyglassFov(Camera camera, float partialTick, boolean optical, CallbackInfoReturnable<Float> cir) {
+    private void modifySpyglassFov(Camera camera, float partialTick, boolean optical, CallbackInfoReturnable<Double> cir) {
         Minecraft client = Minecraft.getInstance();
         if (client.player != null && client.player.isScoping() && client.screen == null) {
-            float fov = cir.getReturnValue();
-            cir.setReturnValue(fov * (float) (10.0 / SpyglassZoom.get(partialTick)));
+            double fov = cir.getReturnValue();
+            cir.setReturnValue(fov * (10.0 / SpyglassZoom.get(partialTick)));
         }
     }
 }
